src/requirements.txt
apiflask
flask_httpauth